C-Obfuscator/requirements.txt
pycparser
pyyaml
pydantic
pydantic_yaml